Move main core and iot header up one directory level.
