+++ sdk/inc/azure/az_core.h
