+++ sdk/inc/azure/az_iot.h
